drm: samsung: run dp_reg_set_aux_pn() depending on usb crossbar support
Type-C crossbar support is enabled by the max77779 chip, which
changes the way SBU polarity is determined on the port. The
AUX channels get inverted by dp_reg_set_aux_pn(), which leads
to the wrong AUX pin being written to the AUX_CONTROL register.
As a result link training fails to read the AUX and DisplayPort
mirroring only works in the normal orientation.

Query the device tree for aux-orientation="auto" during the
driver probe to indicate whether or not the chipset supports
SBU crossbar support.

Add aux_auto_orientation to the DP PHY dp_hw_config struct.
This is used in dp_reg_set_aux_pn() to determine if a
write to AUX_CONTROL is necessary.

Test: manual test on device
Bug: 305311286
Change-Id: I7e000b1abebc7bb360920df6c70614df39d8a5cc
Signed-off-by: RD Babiera <rdbabiera@google.com>