drm: samsung: support set operation rate to min BTS requirement
When the DDIC is in HS mode, its operation rate will be high. If DPU
start to write data early, we can reduce the BTS requirement. If we do
not do this, we need to set a high BTS. Add support to set the
operation rate as the minimum BTS requirement if needed.

Bug: 305987242
Test: NS/HS switch, AoD in/out, in/off, dump dpu events
Change-Id: Ib3a7ef125292b705f5bf2389e33f3138d34187d9
Signed-off-by: Midas Chien <midaschieh@google.com>