display: dsim: Add hs_clock allow table.
From device tree load an allowed list of mipi clock values
to drm_dsim, and added a flag in debugfs entry to enable force
setting clock.
If the force_set_clk flag is off and the list is loaded, only the
specified value in list are allowed to be set from the hs_clock
sysfs node.

Test: with patch pa/2867701, write in allowed/disallowed value
in with force_set_clk on/off then check results and logging.
Bug: 341198690

Change-Id: Iacf8056d763e8d22dec14b17914cb2afed02340b
Signed-off-by: attis <attis@google.com>